improve common syntax highlighter
