+++ app/src/main/assets/highlight/languages/map.properties
@@ -1,0 +1,8 @@
+# C/C++
+c=cpp
+cpp=cpp
+# Java
+java=java
+# Python
+python=python
+py=python
