--- app/src/main/assets/languages/java.json
@@ -1,15 +1,0 @@
-{
-  "language": "Java",
-  "rules": [
-    {
-      "type": "KEYWORD",
-      "regex": "\\b(public|static|void|class)\\b",
-      "color": "#FF5722"
-    },
-    {
-      "type": "STRING",
-      "regex": "\".*\"",
-      "color": "#4CAF50"
-    }
-  ]
-}
