--- app/src/main/assets/languages/map.properties
@@ -1,5 +1,0 @@
-java=java
-.java=java
-.py=python
-markdown=markdown
-.md=markdown
