+++ app/src/main/java/net/gsantner/markor/format/plaintext/PlaintextSyntaxHighlighter.java
@@ -7,26 +7,34 @@ package net.gsantner.markor.format.plaintext;
  *
 #########################################################*/
 
-import net.gsantner.markor.format.plaintext.highlight.Rule;
+import android.graphics.Paint;
+
+import net.gsantner.markor.format.plaintext.highlight.HighlightConfigLoader;
+import net.gsantner.markor.format.plaintext.highlight.Style;
 import net.gsantner.markor.format.plaintext.highlight.Syntax;
-import net.gsantner.markor.format.plaintext.highlight.SyntaxLoader;
 import net.gsantner.markor.frontend.textview.SyntaxHighlighterBase;
 import net.gsantner.markor.model.AppSettings;
 
 import java.util.ArrayList;
 
 public class PlaintextSyntaxHighlighter extends SyntaxHighlighterBase {
-    private final SyntaxLoader syntaxLoader = new SyntaxLoader();
+    public final static HighlightConfigLoader configLoader = new HighlightConfigLoader();
     private Syntax syntax;
+    private Style style;
 
-    public PlaintextSyntaxHighlighter(AppSettings as) {
-        super(as);
+    public PlaintextSyntaxHighlighter(AppSettings appSettings) {
+        super(appSettings);
     }
 
-    public PlaintextSyntaxHighlighter(AppSettings appSettings, String lang) {
+    public PlaintextSyntaxHighlighter(AppSettings appSettings, String extension) {
         super(appSettings);
-        appSettings.getContext();
-        syntax = syntaxLoader.getSyntax(appSettings.getContext(), lang);
+        syntax = configLoader.getSyntax(extension, appSettings.getContext());
+        style = configLoader.getStyle(appSettings.getContext(), "default");
+    }
+
+    @Override
+    public SyntaxHighlighterBase configure(Paint paint) {
+        return super.configure(paint);
     }
 
     @Override
@@ -39,9 +47,14 @@ public class PlaintextSyntaxHighlighter extends SyntaxHighlighterBase {
             return;
         }
 
-        ArrayList<Rule> rules = syntax.getRules();
-        for (Rule rule : rules) {
-            createColorSpanForMatches(rule.getPattern(), rule.getParsedColor());
+        ArrayList<Syntax.Rule> rules = syntax.getRules();
+        ArrayList<Style.Define> defines = style.getDefines();
+        for (Syntax.Rule rule : rules) {
+            for (Style.Define define : defines) {
+                if (define.getType().equals(rule.getType())) {
+                    createColorSpanForMatches(rule.getPattern(), define.getColor());
+                }
+            }
         }
     }
