+++ app/src/main/java/net/gsantner/markor/format/plaintext/highlight/HighlightConfigLoader.java
@@ -1,0 +1,121 @@
+package net.gsantner.markor.format.plaintext.highlight;
+
+import android.content.Context;
+
+import com.google.gson.Gson;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.HashMap;
+import java.util.Properties;
+
+public class HighlightConfigLoader {
+    private final Gson gson = new Gson();
+    private final Properties map = new Properties();
+    private final SyntaxCache syntaxCache = new SyntaxCache();
+    private Style style;
+
+    public static final String MAP_PATH = "highlight/languages/map.properties";
+
+    private void load(Context context, String lang) {
+        try (InputStream input = context.getAssets().open("highlight/languages/" + lang + ".json");) {
+            Syntax syntax = gson.fromJson(new InputStreamReader(input), Syntax.class);
+            syntaxCache.putSyntax(lang, syntax);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void loadStyle(Context context, String name) {
+        try (InputStream input = context.getAssets().open("highlight/styles/" + name + ".json");) {
+            style = gson.fromJson(new InputStreamReader(input), Style.class);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Get language syntax.
+     *
+     * @param context Android Context.
+     * @param lang    Language name or file extension without dot.
+     * @return
+     */
+    public Syntax getSyntax(Context context, String lang) {
+        if (map.isEmpty()) {
+            try {
+                map.load(context.getAssets().open(MAP_PATH));
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        String key = map.getProperty(lang.toLowerCase());
+        Syntax syntax = syntaxCache.getSyntax(key);
+        if (syntax == null) {
+            load(context, key);
+        }
+
+        return syntaxCache.getSyntax(key);
+    }
+
+    public Syntax getSyntax(String extension, Context context) {
+        if (extension.startsWith(".")) {
+            return getSyntax(context, extension.substring(1));
+        } else {
+            return getSyntax(context, extension);
+        }
+    }
+
+    public Style getStyle(Context context, String name) {
+        if (style == null) {
+            loadStyle(context, name);
+        }
+        return style;
+    }
+
+    class SyntaxCache {
+        public static final int CACHE_SIZE = 5;
+        private final HashMap<String, Syntax> syntaxMap = new HashMap<>();
+        private final HashMap<String, Integer> usageMap = new HashMap<>();
+
+        public Syntax getSyntax(String key) {
+            Syntax syntax = syntaxMap.get(key);
+            if (syntax == null) {
+                return null;
+            }
+
+            Integer usage = usageMap.get(key);
+            if (usage == null) {
+                usageMap.put(key, 1);
+            } else {
+                usageMap.put(key, ++usage);
+            }
+
+            return syntax;
+        }
+
+        public void putSyntax(String key, Syntax syntax) {
+            if (syntaxMap.size() > CACHE_SIZE) {
+                String entryKey = null;
+                int min = Integer.MAX_VALUE;
+                for (HashMap.Entry<String, Integer> entry : usageMap.entrySet()) {
+                    int value = entry.getValue();
+                    if (min < value) {
+                        min = value;
+                        entryKey = entry.getKey();
+                    }
+                }
+
+                if (entryKey != null) {
+                    syntaxMap.remove(entryKey);
+                    usageMap.remove(entryKey);
+                }
+            }
+
+            syntaxMap.put(key, syntax);
+            usageMap.put(key, 0);
+        }
+    }
+}
